// src/types/replicants/talent.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Talent = {
	id: string;
	externalId?: string | null;
	name: string;
	pronouns?: string | null;
	countryCode?: string | null;
	socials: {
		twitch?: string | null;
		youtube?: string | null;
		twitter?: string | null;
		speedruncom?: string | null;
		[k: string]: string | null;
	};
}[];